Refactor some folders
