+++ danger_plugins/protect_files.rb
