+++ lib/danger.rb
@@ -1,9 +1,9 @@
 require "danger/version"
-require "danger/dangerfile"
-require "danger/environment_manager"
+require "danger/danger_core/dangerfile"
+require "danger/danger_core/environment_manager"
 require "danger/commands/runner"
-require "danger/available_values"
-require "danger/plugin"
+require "danger/danger_core/available_values"
+require "danger/plugin_support/plugin"
 require "danger/core_ext/string"
 
 require "claide"
