+++ lib/danger/danger_core/available_values.rb
