+++ lib/danger/danger_core/environment_manager.rb
