+++ lib/danger/plugin_support/plugin.rb
